feat(game-round-time): implement round time and cast time display components
